Refactor, moving HTTP stuff out of LcmTestHelper
LcmTestHelper was the wrong place to put CommitAndPush since that needs
to deal with URLs, and URLs are already being dealt with in SRTestEnv.
