+++ src/LfMerge.Core.Tests/LcmTestHelper.cs
@@ -15,58 +15,6 @@ namespace LfMerge.Core.Tests
 {
 	public static class LcmTestHelper
 	{
-		public static string HgCommand =>
-			Path.Combine(TestEnvironment.FindGitRepoRoot(), "Mercurial",
-				Platform.IsWindows ? "hg.exe" : "hg");
-
-		public static string LexboxHostname = Environment.GetEnvironmentVariable(MagicStrings.EnvVar_LanguageDepotPublicHostname) ?? "localhost";
-		public static string LexboxProtocol = Environment.GetEnvironmentVariable(MagicStrings.EnvVar_LanguageDepotUriProtocol) ?? "http";
-		public static string LexboxPort = Environment.GetEnvironmentVariable(MagicStrings.EnvVar_LanguageDepotUriPort) ?? (LexboxProtocol == "http" ? "80" : "443");
-		public static Uri LexboxUrl = new Uri($"{LexboxProtocol}://{LexboxHostname}:{LexboxPort}");
-
-		public static HttpClientHandler Handler { get; set; } = new HttpClientHandler();
-		public static CookieContainer Cookies => Handler.CookieContainer;
-		public static HttpClient Http { get; set; } = new HttpClient(Handler);
-		public static IProgress NullProgress = new NullProgress();
-
-		public static async Task<string> LexboxLogin(string username, string password)
-		{
-			await Http.PostAsJsonAsync(new Uri(LexboxUrl, "/api/login"), new { EmailOrUsername=username, Password=password });
-			var cookies = Cookies.GetCookies(LexboxUrl);
-			return cookies[".LexBoxAuth"].Value;
-		}
-
-		public static FwProject CloneFromLexbox(string code, string baseDir, string? newCode = null)
-		{
-			var projUrl = new Uri(LexboxUrl, $"/hg/{code}");
-			var withAuth = new UriBuilder(projUrl) { UserName = "admin", Password = "pass" };
-			newCode ??= code;
-			var dest = Path.Combine(baseDir, "webwork", newCode);
-			MercurialTestHelper.CloneRepo(withAuth.Uri.AbsoluteUri, dest);
-			var fwdataPath = Path.Join(dest, $"{newCode}.fwdata");
-			var progress = new NullProgress();
-			MercurialTestHelper.ChangeBranch(dest, "tip");
-			LfMergeBridge.LfMergeBridge.ReassembleFwdataFile(progress, false, fwdataPath);
-			var settings = new LfMergeSettingsDouble(baseDir);
-			return new FwProject(settings, newCode);
-		}
-
-		public static void CommitAndPush(FwProject project, string code, string baseDir, string? localCode = null, string? commitMsg = null)
-		{
-			localCode ??= code;
-			var projUrl = new Uri(LexboxUrl, $"/hg/{code}");
-			var withAuth = new UriBuilder(projUrl) { UserName = "admin", Password = "pass" };
-			project.Cache.ActionHandlerAccessor.Commit();
-			if (!project.IsDisposed) project.Dispose();
-			commitMsg ??= "Auto-commit";
-			var projectDir = Path.Combine(baseDir, "webwork", localCode);
-			var fwdataPath = Path.Join(projectDir, $"{localCode}.fwdata");
-			LfMergeBridge.LfMergeBridge.DisassembleFwdataFile(NullProgress, false, fwdataPath);
-			MercurialTestHelper.HgClean(projectDir); // Ensure ConfigurationSettings, etc., don't get committed
-			MercurialTestHelper.HgCommit(projectDir, commitMsg);
-			MercurialTestHelper.HgPush(projectDir, withAuth.Uri.AbsoluteUri);
-		}
-
 		public static IEnumerable<ILexEntry> GetEntries(FwProject project)
 		{
 			return project?.ServiceLocator?.LanguageProject?.LexDbOA?.Entries ?? [];
